Terrain, Boat, Camera and Movement
I've started on a simple terrain, a few mountainous regions to entice the player, and started them in a cove.

the player movement script is pretty basic, and the boat rotates around the front end, represented by the taller block. I'm not sure if this is correct, maybe it should be at the back?

the camera is a little sensitive, so needs tweaking.