StateMachine for shop management
Since are using more than one NPC to purchase stuff from us, It makes sense to try a SM to manage which NPC we're currently closest to.

This way, The game will load the relevant NPC only.

This portion now only needs a Trigger Exit to be complete, putting the active state to NoShop ("you're out at sea")

This was a package I've taken from my previous weeks work. I've removed the ExampleState().