Half Way to Interaction
I have got most of the logic needed to have the interaction between the player and the NPC

The player moves into the trigger area, and is prompted to Press E. On doing so, the Sell Panel pops up, into which the player can drag an item into the slot, and sell it. The button links to the Purchase item method, which sells the item.

HOWEVER, this is where the issue currently is, it's not selling the object in the slot. This needs fixing